move othello-related files into separate folder
