+++ src/flippy/mode/base.py
@@ -1,4 +1,4 @@
-from flippy.board import Board
+from flippy.othello.board import Board
 from pygame.event import Event
