+++ src/flippy/othello/board.py
